Remove reference to mruby-bindings
mruby-bindings has gone through major backward-incompatible changes.
If future code generation is required for this project, it would need
to be forward-ported to the latest mruby-bindings first. This is
unlikely to be necessarry, since the bulk of the code is finished,
or worth it, since it would basically be a from-scratch restart.